backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
faiss-cpu==1.11.0
Faker==37.4.0
numpy==2.3.1
packaging==25.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
